+++ rules/rule.go
@@ -13,10 +13,11 @@ const (
 	LOAD_ERR1 = "规则名为空"
 	LOAD_ERR2 = "规则(%s)未注册"
 	LOAD_ERR3 = "生成%s规则失败"
+	HEADER    = ">>>>>>>>开始处理<<<<<<<<"
 )
 
 type Rule struct {
-	ruler        Ruler                   `description:"规则"`
+	ruler        []Ruler                 `description:"规则,数组是指统一规则的多个worker"`
 	eventChannel chan event.Event        `description:"事件队列"`
 	errHr        *cobraErrors.ErrHandler `description:"错误处理对象"`
 	isRulerClose chan bool
@@ -24,6 +25,8 @@ type Rule struct {
 	isReady      bool
 	Log          *log.Logger
 	rulerNum     int
+	aggregator   config.Aggregatable
+	name         string
 }
 
 var ruleMakers = map[string]func(config.RuleConfig) (Ruler, error){
@@ -45,7 +48,6 @@ func CreateRule(cfg config.RuleConfig) (rule Rule, err error) {
 	if rule, err = InitRule(cfg); err != nil {
 		return
 	}
-	var ruler Ruler
 	if cfg.Name == "" {
 		rule.Log.Info("构建fake规则......")
 		cfg.Name = "fake"
@@ -54,16 +56,25 @@ func CreateRule(cfg config.RuleConfig) (rule Rule, err error) {
 	if !ok {
 		err = errors.Errorf(LOAD_ERR2, cfg.Name)
 	}
-	ruler, err = f(cfg)
-	if ruler == nil {
-		err = errors.Errorf(LOAD_ERR3, cfg.Name)
-		return
+	for i := 0; i < rule.rulerNum; i++ {
+		var ruler Ruler
+		if ruler, err = f(cfg); err != nil {
+			return
+		}
+		if ruler == nil {
+			err = errors.Errorf(LOAD_ERR3, cfg.Name)
+			return
+		}
+		ruler.SetNumber(i)
+		ruler.SetLogger(rule.Log)
+		if err = ruler.LoadConfig(cfg); err != nil {
+			return
+		}
+		if rule.IsAggre() {
+			ruler.SetAggregator(rule.aggregator)
+		}
+		rule.SetRuler(ruler)
 	}
-	ruler.SetLogger(rule.Log)
-	if err = ruler.LoadConfig(cfg); err != nil {
-		return
-	}
-	rule.SetRuler(ruler)
 	return
 }
 
@@ -74,16 +85,33 @@ func InitRule(cfg config.RuleConfig) (rule Rule, err error) {
 	rule.isRulerClose = make(chan bool, 1)
 	rule.Log, err = cfg.LogCfg.GetLogger()
 	rule.rulerNum = cfg.Worker()
+	rule.aggregator = cfg.InitAggregator()
+	rule.name = cfg.Name
 	return
 }
 
+func (this *Rule) SetName(name string) {
+	this.name = name
+}
+
+func (this *Rule) GetName() string {
+	return this.name
+}
+
+func (this *Rule) IsAggre() bool {
+	return this.aggregator != nil
+}
+
 func (this *Rule) SetRuler(r Ruler) {
-	this.ruler = r
+	if this.ruler == nil {
+		this.ruler = make([]Ruler, 0)
+	}
+	this.ruler = append(this.ruler, r)
 }
 
 func (this *Rule) Push(e event.Event) {
 	if this.closed {
-		this.Log.Errorf("%s规则事件池已经关闭，不能放入事件", this.ruler.GetName())
+		this.Log.Errorf("%s规则事件池已经关闭，不能放入事件", this.name)
 		return
 	}
 	this.eventChannel <- e
@@ -96,9 +124,19 @@ func (this *Rule) Close() error {
 	close(this.eventChannel)
 	this.closed = true
 	<-this.isRulerClose
-	err := this.ruler.Close()
+	var err error
+	for _, ruler := range this.ruler {
+		e := ruler.Close()
+		if e != nil {
+			if err != nil {
+				err = errors.Errorf("%s->%s", err, e)
+			} else {
+				err = e
+			}
+		}
+	}
 	this.errHr.Close()
-	this.Log.Infof("%s规则的错误处理器关闭", this.ruler.GetName())
+	this.Log.Infof("%s规则的错误处理器关闭", this.name)
 	return err
 }
 
@@ -106,43 +144,35 @@ func (this *Rule) Start() {
 	if this.isReady {
 		return
 	}
-	this.Log.Infof("%s规则的事件池开启...", this.ruler.GetName())
+	this.Log.Infof("%s规则的事件池开启...", this.name)
 	this.isReady = true
 	go this.errHr.Send()
-	this.Log.Infof("%s规则的错误处理器开启", this.ruler.GetName())
-	this.ruler.Start()
+	this.Log.Infof("%s规则的错误处理器开启", this.name)
 	var wg sync.WaitGroup
-	for i := 0; i < this.rulerNum; i++ {
+	for _, ruler := range this.ruler {
+		ruler.Start()
 		wg.Add(1)
-		go func(num int) {
+		go func(r Ruler) {
 			for {
 				e, isOpen := <-this.eventChannel
 				if !isOpen {
 					break
 				}
-				this.Log.Debugf("Worker%d: %s规则发现有事件需要处理:\n%s", num, this.ruler.GetName(), e.String())
-				if err := this.ruler.HandleEvent(e, num); err != nil {
+				this.Log.Debugf("Rule%d: %s", r.GetNumber(), HEADER)
+				this.Log.Debugf("Rule%d: %s规则发现有事件需要处理:%s", r.GetNumber(), this.name, e.String())
+				if err := r.HandleEvent(e); err != nil {
 					go this.errHr.Push(err)
 				}
-				this.Log.Debug("Worker%d: 处理完毕", num)
+				this.Log.Debugf("Rule%d: 处理完毕", r.GetNumber())
 			}
 			wg.Done()
-		}(i)
+		}(ruler)
 	}
 	wg.Wait()
-	/*
-		for {
-			e, isOpen := <-this.eventChannel
-			if !isOpen {
-				break
-			}
-			this.Log.Debugf("%s规则发现有事件需要处理:\n%s", this.ruler.GetName(), e.String())
-			if err := this.ruler.HandleEvent(e); err != nil {
-				go this.errHr.Push(err)
-			}
-			this.Log.Debug("处理完毕")
-		}
-	*/
-	this.Log.Infof("%s规则的事件池关闭", this.ruler.GetName())
+	this.Log.Infof("%s规则的事件池关闭", this.name)
+	if this.IsAggre() {
+		this.aggregator.Stop()
+		this.Log.Infof("%s规则关闭聚合器", this.name)
+	}
 	this.isRulerClose <- true
 }