+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - event增加解析函数
 - rule增加新的mysql链接
+- 把rule变为可以有多个处理的模式，用worker确认
+- 把consume变为可以有多个处理的模式，用worker确认
 
 ## [1.0.2] - 2019-12-06
 ### Changed